pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "llama-cloud-services>=0.6.1",
    "llama-index-core>=0.12.19",
    "llama-index-readers-file>=0.4.5",
    "marker-pdf>=1.5.5",
    "python-dotenv>=1.0.1",
]
